+++ src/widgets/stage/create/official/constants/matchTypes.ts
@@ -1,0 +1,5 @@
+export const matchTypeOptions = [
+  { value: 'TOURNAMENT', label: '토너먼트' },
+  { value: 'FULL_LEAGUE', label: '리그전' },
+  { value: 'SINGLE', label: '단판' },
+];
