+++ src/widgets/stage/create/official/ui/MatchSettingContainer/index.tsx
@@ -1,17 +1,13 @@
-import React, { useState } from 'react';
-import {
-  useFieldArray,
-  UseFormRegister,
-  Control,
-  UseFormWatch,
-} from 'react-hook-form';
-import { toast } from 'react-toastify';
+import React from 'react';
+import { UseFormRegister, Control, UseFormWatch } from 'react-hook-form';
 import { AddButton } from '@/entities/stage/create/official';
 import GameField from '@/entities/stage/create/official/ui/GameField';
-import { SportType } from '@/shared/model/sportTypes';
 import { OfficialStageData } from '@/shared/types/stage/create/official';
 import SportTypeLabel from '@/shared/ui/sportTypelabel';
 import { cn } from '@/shared/utils/cn';
+import { matchTypeOptions } from '../../constants/matchTypes';
+import { categoryTypes } from '../../constants/sportTypes';
+import { useMatchSetting } from '../../model/useMatchSetting';
 
 interface MatchSettingContainerProps {
   control: Control<OfficialStageData>;
@@ -24,94 +20,14 @@ const MatchSettingContainer = ({
   register,
   watch,
 }: MatchSettingContainerProps) => {
-  const { fields, append, remove } = useFieldArray({
-    control,
-    name: 'game',
-  });
-
-  const [selectedSport, setSelectedSport] = useState<SportType | null>(null);
-
-  const categoryTypes: SportType[] = [
-    'VOLLEY_BALL',
-    'SOCCER',
-    'LOL',
-    'BASE_BALL',
-    'BASKET_BALL',
-    'BADMINTON',
-    'ETC',
-  ];
-
-  const matchTypeOptions = [
-    { value: 'TOURNAMENT', label: '토너먼트' },
-    { value: 'FULL_LEAGUE', label: '리그전' },
-    { value: 'SINGLE', label: '단판' },
-  ];
-
-  const getCategoryLabel = (category: string) => {
-    const categoryMap: { [key: string]: string } = {
-      VOLLEY_BALL: '배구',
-      SOCCER: '축구',
-      LOL: 'LOL',
-      BASE_BALL: '야구',
-      BASKET_BALL: '농구',
-      BADMINTON: '배드민턴',
-      ETC: '기타',
-    };
-    return categoryMap[category] || category;
-  };
-
-  const validateCurrentSportFields = () => {
-    if (!selectedSport) return true;
-
-    const currentFields = fields.filter(
-      (field) => field.category === selectedSport,
-    );
-    if (currentFields.length === 0) return true;
-
-    for (let i = 0; i < currentFields.length; i++) {
-      const fieldIndex = fields.findIndex(
-        (field) => field.id === currentFields[i].id,
-      );
-      const fieldName = watch(`game.${fieldIndex}.name`);
-      const fieldMinCapacity = watch(`game.${fieldIndex}.teamMinCapacity`);
-      const fieldMaxCapacity = watch(`game.${fieldIndex}.teamMaxCapacity`);
-
-      if (!fieldName || !fieldMinCapacity || !fieldMaxCapacity) {
-        return false;
-      }
-    }
-
-    return true;
-  };
-
-  const handleSportSelect = (sport: SportType) => {
-    if (selectedSport && selectedSport !== sport) {
-      if (!validateCurrentSportFields()) {
-        toast.error(
-          `${getCategoryLabel(selectedSport)} 카테고리의 모든 필드를 작성해주세요.`,
-        );
-        return;
-      }
-    }
-
-    setSelectedSport(sport);
-  };
-
-  const handleAddGame = () => {
-    if (selectedSport) {
-      append({
-        category: selectedSport,
-        name: '',
-        system: 'TOURNAMENT',
-        teamMinCapacity: null,
-        teamMaxCapacity: null,
-      });
-    }
-  };
-
-  const filteredFields = fields.filter((field) =>
-    selectedSport ? field.category.includes(selectedSport) : true,
-  );
+  const {
+    selectedSport,
+    fields,
+    filteredFields,
+    remove,
+    handleSportSelect,
+    handleAddGame,
+  } = useMatchSetting(control, watch);
 
   return (
     <div className={cn('space-y-16')}>
